requirements.txt
flask
python-dotenv
langchain
langchain-groq
flask-socketio
eventlet
pydantic
pydantic[email]
mysql-connector-python